Put the event actions in their own file
